Modified fix_load_segments.repair to take a list of ifot_loads instead of a db handle.
This way, an in-memory list of loads may be patched, instead of
inserting the list into the db and then fixing the db.  The previous
method of inserting the list into the db and then modifying it is
problematic as a cron task as it doesn't work in sybase due to foreign
key constraints.  And it was clunky anyway.

Testing data was added and updated to work with new cases and new
nonload_cmds.