Updated the "with-gratings" fiducial data again.
Due to some trouble with column order (hetg/letg in the middle or at
the end of the table) redid string_states to put out text tables with
fields alphebetized and updated fiducial data to match.

Also updated the 2010 NSM test to use a subdirectory of t/ for its
output like the rest of the suite.