Updated fiducial data for Chandra.Maneuver 0.04
Chandra.Maneuver 0.04 adds a fix to attitudes() to set the final attitude
to be exactly the requested final attitude/quaternion.  This causes
tiny changes in ra/dec/roll/pitch in cmd_states, and as such the
fiducial data files of text files of cmd_states needed to be updated
to reflect the changes.

Additionally, timelines_make_testdb.py was fixed to reflect the
current API of populate_states, and a duplicate test was removed from
timelines_test.py.