Fixed tiny bug in fix_load_segments.py that had a large effect on update_load_seg_db.py if run on a load_rdb including more than one calendar year.
fix_load_segments.py was performing a sort on the load segment id
(e.g. 'CL110:0205') after making any changes.  This fails if the tool
is passed a list of loads that crosses a year boundary, as the
segments from the later year will restart at CL001.

Fixed by performing the sort on the id (int(tstart)).

Updated the version number in conf.py.

Modified test_nsm_2010 in timelines_test.py to let it "succeed" if it
had trouble deleting the directory (not sure why, but it seems to get
hung up on .nfs lock files in that directory a lot).