Modified load update code to not remove last good load.
All of update_load_seg_db.py is now more aware of load_scs and
timelines and load segments are sorted by that secondarily for more
reliable comparisons.  The difference determination section for loads
now sets the i_diff after the good loads in the case of a pure append
to the loads so a last good load segment will not be deleted.  The
regression tests designed to test this have had their bugs fixed.