Change cmd_state copy over
The tests can only work to build cmd_states from a full cmd_state
and complete cmds.  cmd_states need to be copied over past the start
of the interval over which there are timelines.  This change copies
up through the first NPNT which has a *not* NPNT after it.  In the
case of sampled states for pitch, this should get through the whole
dwell.